Start restructuring reports
The reports are now generated according to functions in a Report
trait. This is split into rendering (creating a type which represents
serialisation of traces and config into a report) and export (uploading
or saving this report).

The coveralls and cobertura modules have been redefined to use this
new style. The cobertura module is currently half complete, as it
does not currently export the generated type as XML. Some parts
are also not rendered properly, as branch coverage has yet to be
impemented in the main library.