Move stdout reporting into reporting module
The reporting module now has stdout as another reporting backend.
This means the explicit stdout reporting can be changed / removed
without needing to dig into the reporting function in lib.rs.

This also means that if a quiet mode is needed, it is now super
easy to print nothing to stdout at all.